station: add {Set,Unset}ConnectedBssAffinity DBus methods
These methods can be used to loosly lock IWD to its current BSS.
Calling SetConnectedBssAffinity() will lower the roaming threshold
to a critical level ([General].CriticalRoamThreshold) thereby
preventing IWD from trying to roam unless the RSSI reaches this
level.

The motivation for this is a use case in which an external
client/daemon has additional information about the environment and
can use this to make an informed decision about roaming. For example
if the device is stationary its unlikely IWD will find a better roam
target.