network: implement net.connman.iwd.BasicServiceSet
This ties the addition/removal of scan_bss objects from a network
object to BasicServiceSet objects. If a "new" object is added when
one with the same address already existed, no DBus object changes
are made (only the user data is updated with the new scan_bss).

Currently the only property on a BasicServiceSet object is the
"Bssid".