doc: Document station Affinities property
This documents new DBus property that expose a bit more control to
how IWD roams.

Setting the affinity on the connected BSS effectively "locks" IWD to
that BSS (except at critical RSSI levels, explained below). This can
be useful for clients that have access to more information about the
environment than IWD. For example, if a client is stationary there
is likely no point in trying to roam until it has moved elsewhere.

Setting the affinity on a BSS (regardless if the connected BSS) will
also cause a modified ranking factor to be applied to the BSS. This
could be used to push IWD to certain BSS's over others. Similar to
above, the affinity ranking factor for a BSS will only be applied if
that BSS's signal is above the crtical roaming threshold.

A new main.conf option would also be added:

[General].CriticalRoamThreshold

This would be the new roam threshold set if the currently connected
BSS is in the Affinities list. If the RSSI continues to drop below
this level IWD will still attempt to roam.