station: add Affinities DBus property
This property will hold an array of object paths for
BasicServiceSet (BSS) objects. For the purpose of this patch
only the setter/getter and client watch is implemented. The
purpose of this array is to guide or loosly lock IWD to certain
BSS's provided that some external client has more information
about the environment than what IWD takes into account for its
roaming decisions. For example:

 - If the device is stationary an external client could reduce
   the likelihood of roaming by setting the affinity to the
   current BSS.
 - If an external client has knowledge of the devices planned
   route though the environment it could plan a path to roam
   between particular APs to reduce and optimize roaming.