station: Use Affinities property for roaming/ranking decisions
There are two pieces where the affinity to a BSS will come into play:

The first is while connected to a BSS with the affinity set. In this
scenario station will lower the roaming threshold in order to loosly
lock IWD to this BSS. By lowering the roaming threshold we can avoid
roaming/scanning in cases where the device is not moving i.e. if an
external client knows this information.

The second effect BSS affinity has is to modify ranking when choosing
a roam candidate. A new ranking modifier has been added that will be
applied to a BSS rank given:
 - The BSS is in the Affinities array
 - The BSS signal is above the critical threshold (set in main.conf
   [General].CriticalRoamThreshold{_5GHZ}).