netdev: loosen parsing requirements on STA_INFO
If one of the nested attributes in STA_INFO isn't exactly what we
expect in terms of length or content skip it rather than fatally
return. Each of the diagnostic info members uses a bool to indicate
if it was found, so if one is unable to parse it just won't be
present in the diagnostic message.