netdev: set a timeout for CMD_CONNECT based connections
Based on a users report/logs it appears some drivers (brcmfmac in
this case) may have issues which cause them to never respond with
a CMD_CONNECT event, leaving IWD hung waiting indefinitely for
this.

Since IWD cannot control driver behavior its safest to set a timer
in userspace as a last resort in case this happens.