station: improve roam scan strategy
When IWD decides to roam, it scans either neighbor freqs or known freqs
(if neighbors are not available). If it fails to roam after getting the
scan results, it scans ALL freqs. In my testing there's a high chance
that both neighbor and/or known scans fail to roam and we end up
scanning all freqs. This is very slow and if you're already moving away
from the current BSS, there's a high chance you will lose connection
completely before the scan is finished.

Instead of scanning all freqs at once, split them up into prioritized
subsets. Each subset contains a handful of freqs each, a lower index for
the subset means that its freqs are more common. So subset 0 has the
most common freqs and subset 3 has the least common freqs. The first two
subsets also contain no DFS channels, speeding up scanning even more. In
order to make this efficient, use the "scan_freq_map" to avoid scanning
freqs that were recently scanned.

When a roam scan is triggered, add the most prioritized freqs to the
list of freqs that should be scanned. The order of priority is:

1. Neighbor freqs
2. Known freqs
3. Subsets, starting with index 0 and incrementing if the subset is
   exhausted

For each freq candidate, check the scan_freq_map to see how long time
ago this freq was last scanned, this is denoted as its "age". If its age
is above a threshold, add the freq to the list, otherwise discard it.
Once the list has a certain size, start the scan. If no roaming occurs
after the scan is completed, run the same function again. It will now
pick new freqs since the previous freqs have been updated in
scan_freq_map.

This approach results in more scans, but fewer freqs per scan, leading
to shorter delays between scan results. It also avoids scanning the same
freqs back-to-back, which is generally not very useful. In combination
with the freq priority, this increases the chance of finding a good BSS
early.